Modularize rwlock
