+++ c-scape/src/thread/mod.rs
@@ -1,4 +1,5 @@
 mod key;
+mod rwlock;
 mod spinlock;
 
 use crate::GetThreadId;
@@ -9,12 +10,12 @@ use core::mem::{align_of, size_of, transmute, zeroed, ManuallyDrop, MaybeUninit}
 use core::ptr::{self, copy_nonoverlapping, null_mut, NonNull};
 use core::slice;
 use core::sync::atomic::Ordering::SeqCst;
-use core::sync::atomic::{AtomicBool, AtomicU32};
+use core::sync::atomic::AtomicU32;
 use core::time::Duration;
 use origin::thread::{self, Thread};
 use rustix::fs::{Mode, OFlags};
-use rustix_futex_sync::lock_api::{RawMutex as _, RawReentrantMutex, RawRwLock as _};
-use rustix_futex_sync::{Once, RawCondvar, RawMutex, RawRwLock};
+use rustix_futex_sync::lock_api::{RawMutex as _, RawReentrantMutex};
+use rustix_futex_sync::{Once, RawCondvar, RawMutex};
 
 use libc::{c_char, c_int, size_t};
 
@@ -90,18 +91,6 @@ struct PthreadMutexT {
 }
 libc_type!(PthreadMutexT, pthread_mutex_t);
 
-#[allow(non_camel_case_types)]
-#[repr(C)]
-struct PthreadRwlockT {
-    lock: RawRwLock,
-    exclusive: AtomicBool,
-    pad0: usize,
-    pad1: usize,
-    pad2: usize,
-    pad3: usize,
-    pad4: usize,
-}
-libc_type!(PthreadRwlockT, pthread_rwlock_t);
 
 #[allow(non_camel_case_types)]
 #[repr(C)]
@@ -111,15 +100,6 @@ struct PthreadMutexattrT {
     pad0: u32,
 }
 libc_type!(PthreadMutexattrT, pthread_mutexattr_t);
-
-#[allow(non_camel_case_types)]
-#[cfg_attr(target_pointer_width = "32", repr(C, align(4)))]
-#[cfg_attr(target_pointer_width = "64", repr(C, align(8)))]
-struct PthreadRwlockattrT {
-    kind: AtomicU32,
-    pad0: u32,
-}
-libc_type!(PthreadRwlockattrT, pthread_rwlockattr_t);
 
 #[no_mangle]
 unsafe extern "C" fn pthread_self() -> PthreadT {
@@ -375,57 +355,6 @@ unsafe extern "C" fn pthread_mutex_unlock(mutex: *mut PthreadMutexT) -> c_int {
 }
 
 #[no_mangle]
-unsafe extern "C" fn pthread_rwlock_tryrdlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_tryrdlock(checked_cast!(rwlock)));
-    let result = (*rwlock).lock.try_lock_shared();
-    if result {
-        (*rwlock).exclusive.store(false, SeqCst);
-        0
-    } else {
-        libc::EBUSY
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_trywrlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_trywrlock(checked_cast!(rwlock)));
-    let result = (*rwlock).lock.try_lock_exclusive();
-    if result {
-        (*rwlock).exclusive.store(true, SeqCst);
-        0
-    } else {
-        libc::EBUSY
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_rdlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_rdlock(checked_cast!(rwlock)));
-    (*rwlock).lock.lock_shared();
-    (*rwlock).exclusive.store(false, SeqCst);
-    0
-}
-
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_unlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_unlock(checked_cast!(rwlock)));
-
-    let rwlock = &*rwlock;
-    if rwlock.exclusive.load(SeqCst) {
-        if !rwlock.lock.is_locked_exclusive() {
-            return libc::EPERM;
-        }
-        rwlock.lock.unlock_exclusive();
-    } else {
-        if !rwlock.lock.is_locked() {
-            return libc::EPERM;
-        }
-        rwlock.lock.unlock_shared();
-    }
-    0
-}
-
-#[no_mangle]
 unsafe extern "C" fn pthread_attr_getguardsize(
     attr: *const PthreadAttrT,
     guardsize: *mut usize,
@@ -617,53 +546,6 @@ unsafe extern "C" fn pthread_cond_timedwait(
         libc::PTHREAD_MUTEX_ERRORCHECK => todo!("PTHREAD_MUTEX_ERRORCHECK"),
         other => unimplemented!("unsupported pthread mutex kind {}", other),
     }
-}
-
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_init(
-    rwlock: *mut PthreadRwlockT,
-    rwlockattr: *const PthreadRwlockattrT,
-) -> c_int {
-    libc!(libc::pthread_rwlock_init(
-        checked_cast!(rwlock),
-        checked_cast!(rwlockattr)
-    ));
-    ptr::write(&mut (*rwlock).lock, RawRwLock::INIT);
-    (*rwlock).exclusive.store(false, SeqCst);
-
-    0
-}
-
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_destroy(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_destroy(checked_cast!(rwlock)));
-    ptr::drop_in_place(rwlock);
-    0
-}
-
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_wrlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_wrlock(checked_cast!(rwlock)));
-    (*rwlock).lock.lock_exclusive();
-    (*rwlock).exclusive.store(true, SeqCst);
-    0
-}
-
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlockattr_init(attr: *mut PthreadRwlockattrT) -> c_int {
-    libc!(libc::pthread_rwlockattr_init(checked_cast!(attr)));
-
-    attr.write(PthreadRwlockattrT {
-        kind: AtomicU32::new(0),
-        pad0: 0,
-    });
-    0
-}
-
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlockattr_destroy(_attr: *mut PthreadRwlockattrT) -> c_int {
-    libc!(libc::pthread_rwlockattr_destroy(checked_cast!(_attr)));
-    0
 }
 
 #[no_mangle]
